chore: restructure files and add reusable component for Header
