+++ components/keyboard/IntroText.tsx
